+++ Sources/NSFetchRequestResult+CoreData.swift
@@ -27,7 +27,7 @@ extension NSFetchRequestResult where Self: NSManagedObject {
   /// **CoreDataPlus**
   ///
   /// The entity name.
-  @available(iOS 10, tvOS 10, watchOS 3, macOS 10.12, *)
+  @available(iOS 10.0, tvOS 10.0, watchOS 3.0, macOS 10.12, *)
   public static var entityName: String {
     if let name = entity().name {
       return name
@@ -40,35 +40,35 @@ extension NSFetchRequestResult where Self: NSManagedObject {
     // https://github.com/jessesquires/rdar-19368054
     return String(describing: Self.self)
   }
-
+  
   // MARK: - Fetch
-
+  
   /// **CoreDataPlus**
   ///
   /// Creates a `new` NSFetchRequest for `self`.
   /// - Note: Use this method instead of fetchRequest() to avoid a bug in CoreData occurring in the Unit Test targets or when Generics are used.
-  @available(iOS 10, tvOS 10, watchOS 3, macOS 10.12, *)
+  @available(iOS 10.0, tvOS 10.0, watchOS 3.0, macOS 10.12, *)
   public static func newFetchRequest() -> NSFetchRequest<Self> {
     let fetchRequest = NSFetchRequest<Self>(entityName: entityName)
     return fetchRequest
   }
-
+  
   /// **CoreDataPlus**
   ///
   /// Performs a configurable fetch request in a context.
   /// - Throws: It throws an error in cases of failure.
-  @available(iOS 10, tvOS 10, watchOS 3, macOS 10.12, *)
+  @available(iOS 10.0, tvOS 10.0, watchOS 3.0, macOS 10.12, *)
   public static func fetch(in context: NSManagedObjectContext, with configuration: (NSFetchRequest<Self>) -> Void = { _ in }) throws -> [Self] {
     let request = NSFetchRequest<Self>(entityName: entityName)
     configuration(request)
-
+    
     do {
       return try context.fetch(request)
     } catch {
       throw NSError.fetchFailed(underlyingError: error)
     }
   }
-
+  
   /// **CoreDataPlus**
   ///
   /// Fetches all the `NSManagedObjectID` for a given predicate.
@@ -86,16 +86,16 @@ extension NSFetchRequestResult where Self: NSManagedObject {
     request.resultType = .managedObjectIDResultType
     request.includesSubentities = includingSubentities
     request.predicate = predicate
-
+    
     do {
       return try context.fetch(request)
     } catch {
       throw NSError.fetchFailed(underlyingError: error)
     }
   }
-
+  
   // MARK: - First
-
+  
   /// **CoreDataPlus**
   ///
   /// Attempts to find an object matching a predicate or creates a new one and configures it (if multiple objects are found, configures the **first** one).
@@ -106,18 +106,18 @@ extension NSFetchRequestResult where Self: NSManagedObject {
   ///   - configuration: Configuration closure called **only** when creating a new object.
   /// - Returns: A matching object or a configured new one.
   /// - Throws: It throws an error in cases of failure.
-  @available(iOS 10, tvOS 10, watchOS 3, macOS 10.12, *)
+  @available(iOS 10.0, tvOS 10.0, watchOS 3.0, macOS 10.12, *)
   public static func findOneOrCreate(in context: NSManagedObjectContext, where predicate: NSPredicate, with configuration: (Self) -> Void) throws -> Self {
     guard let object = try findOneOrFetch(in: context, where: predicate) else {
       let newObject = Self(context: context)
       configuration(newObject)
-
+      
       return newObject
     }
-
+    
     return object
   }
-
+  
   /// **CoreDataPlus**
   ///
   /// Tries to find the first existing object in the context (memory) matching a predicate.
@@ -128,7 +128,7 @@ extension NSFetchRequestResult where Self: NSManagedObject {
   ///   - predicate: Matching predicate.
   /// - Returns: The first matching object (if any).
   /// - Throws: It throws an error in cases of failure.
-  @available(iOS 10, tvOS 10, watchOS 3, macOS 10.12, *)
+  @available(iOS 10.0, tvOS 10.0, watchOS 3.0, macOS 10.12, *)
   public static func findOneOrFetch(in context: NSManagedObjectContext, where predicate: NSPredicate) throws -> Self? {
     // first we should fetch an existing object in the context as a performance optimization
     guard let object = findOneMaterializedObject(in: context, where: predicate) else {
@@ -143,12 +143,12 @@ extension NSFetchRequestResult where Self: NSManagedObject {
         throw NSError.fetchFailed(underlyingError: error)
       }
     }
-
+    
     return object
   }
-
+  
   // MARK: - Unique
-
+  
   /// **CoreDataPlus**
   ///
   /// Attempts to find an unique object matching a predicate or creates a new one and configures it.
@@ -159,18 +159,18 @@ extension NSFetchRequestResult where Self: NSManagedObject {
   ///   - configuration: Configuration closure called **only** when creating a new object.
   /// - Returns: A matching object or a configured new one.
   /// - Throws: It throws an error in cases of failure or if multiple objects are found.
-  @available(iOS 10, tvOS 10, watchOS 3, macOS 10.12, *)
+  @available(iOS 10.0, tvOS 10.0, watchOS 3.0, macOS 10.12, *)
   public static func findUniqueOrCreate(in context: NSManagedObjectContext, where predicate: NSPredicate, with configuration: (Self) -> Void) throws -> Self {
     guard let object = try findUniqueOrFetch(in: context, where: predicate) else {
       let newObject = Self(context: context)
       configuration(newObject)
-
+      
       return newObject
     }
-
+    
     return object
   }
-
+  
   /// **CoreDataPlus**
   ///
   /// Tries to find an unique existing object in the context (memory) matching a predicate.
@@ -181,7 +181,7 @@ extension NSFetchRequestResult where Self: NSManagedObject {
   ///   - predicate: Matching predicate.
   /// - Returns: The first matching object (if any).
   /// - Throws: It throws an error in cases of failure or if multiple objects are found.
-  @available(iOS 10, tvOS 10, watchOS 3, macOS 10.12, *)
+  @available(iOS 10.0, tvOS 10.0, watchOS 3.0, macOS 10.12, *)
   public static func findUniqueOrFetch(in context: NSManagedObjectContext, where predicate: NSPredicate) throws -> Self? {
     guard let object = try findUniqueMaterializedObject(in: context, where: predicate) else {
       do {
@@ -192,10 +192,10 @@ extension NSFetchRequestResult where Self: NSManagedObject {
         throw NSError.fetchFailed(underlyingError: error)
       }
     }
-
+    
     return object
   }
-
+  
   /// **CoreDataPlus**
   ///
   /// Iterates over the context’s registeredObjects set (which contains all managed objects the context currently knows about) until it finds
@@ -209,18 +209,18 @@ extension NSFetchRequestResult where Self: NSManagedObject {
     }
     return results.first
   }
-
+  
   /// **CoreDataPlus**
   ///
   /// Executes a fetch request where **only** a single object is expected as result, otherwhise a an error is thrown.
   /// - Throws: It throws an error in cases of failure.
-  @available(iOS 10, tvOS 10, watchOS 3, macOS 10.12, *)
+  @available(iOS 10.0, tvOS 10.0, watchOS 3.0, macOS 10.12, *)
   public static func fetchUniqueObject(in context: NSManagedObjectContext, with configuration: @escaping (NSFetchRequest<Self>) -> Void) throws -> Self? {
     let result = try fetch(in: context) { request in
       configuration(request)
       request.fetchLimit = 2
     }
-
+    
     switch result.count {
     case 0:
       return nil
@@ -230,9 +230,9 @@ extension NSFetchRequestResult where Self: NSManagedObject {
       throw NSError.fetchExpectingOnlyOneObjectFailed()
     }
   }
-
+  
   // MARK: - Delete
-
+  
   /// **CoreDataPlus**
   ///
   /// Specifies the objects (matching a given predicate) that should be removed from its persistent store when changes are committed.
@@ -240,7 +240,7 @@ extension NSFetchRequestResult where Self: NSManagedObject {
   /// If `includingSubentities` is set to `false`, sub-entities will be ignored.
   /// - Note: `NSBatchDeleteRequest` would be more efficient but requires a context with an `NSPersistentStoreCoordinator` directly connected (no child context).
   /// - Throws: It throws an error in cases of failure.
-  @available(iOS 10, tvOS 10, watchOS 3, macOS 10.12, *)
+  @available(iOS 10.0, tvOS 10.0, watchOS 3.0, macOS 10.12, *)
   public static func deleteAll(in context: NSManagedObjectContext, includingSubentities: Bool = true, where predicate: NSPredicate = NSPredicate(value: true)) throws {
     do {
       try autoreleasepool {
@@ -254,7 +254,7 @@ extension NSFetchRequestResult where Self: NSManagedObject {
       throw NSError.fetchFailed(underlyingError: error)
     }
   }
-
+  
   /// **CoreDataPlus**
   ///
   /// Removes all entities from within the specified `NSManagedObjectContext` excluding a given list of entities.
@@ -264,31 +264,31 @@ extension NSFetchRequestResult where Self: NSManagedObject {
   ///   - objects: An Array of `NSManagedObjects` belonging to the `NSManagedObjectContext` to exclude from deletion.
   /// - Throws: It throws an error in cases of failure.
   /// - Note: `NSBatchDeleteRequest` would be more efficient but requires a context with an `NSPersistentStoreCoordinator` directly connected (no child context).
-  @available(iOS 10, tvOS 10, watchOS 3, macOS 10.12, *)
+  @available(iOS 10.0, tvOS 10.0, watchOS 3.0, macOS 10.12, *)
   public static func deleteAll(in context: NSManagedObjectContext, except objects: [Self]) throws {
     let predicate = NSPredicate(format: "NOT (self IN %@)", objects)
     try deleteAll(in: context, includingSubentities: true, where: predicate )
   }
-
+  
   // MARK: - Count
-
+  
   /// **CoreDataPlus**
   ///
   /// Counts the results of a configurable fetch request in a context.
   /// - Throws: It throws an error in cases of failure.
-  @available(iOS 10, tvOS 10, watchOS 3, macOS 10.12, *)
+  @available(iOS 10.0, tvOS 10.0, watchOS 3.0, macOS 10.12, *)
   public static func count(in context: NSManagedObjectContext, for configuration: (NSFetchRequest<Self>) -> Void = { _ in }) throws -> Int {
     let request = newFetchRequest()
     configuration(request)
-
+    
     let result = try context.count(for: request)
     guard result != NSNotFound else { throw NSError.fetchCountFailed() }
-
+    
     return result
   }
-
+  
   // MARK: - Materialized Object
-
+  
   /// **CoreDataPlus**
   ///
   /// Iterates over the context’s registeredObjects set (which contains all managed objects the context currently knows about) until it finds one that is not a fault matching for a given predicate.
@@ -296,13 +296,13 @@ extension NSFetchRequestResult where Self: NSManagedObject {
   public static func findOneMaterializedObject(in context: NSManagedObjectContext, where predicate: NSPredicate) -> Self? {
     for object in context.registeredObjects where !object.isFault {
       guard let result = object as? Self, predicate.evaluate(with: result) else { continue }
-
+      
       return result
     }
-
+    
     return nil
   }
-
+  
   /// **CoreDataPlus**
   ///
   /// Iterates over the context’s registeredObjects set (which contains all managed objects the context currently knows about) until it finds
@@ -310,7 +310,7 @@ extension NSFetchRequestResult where Self: NSManagedObject {
   /// Faulted objects are not considered to prevent Core Data to make a round trip to the persistent store.
   public static func findMaterializedObjects(in context: NSManagedObjectContext, where predicate: NSPredicate) -> [Self] {
     let results = context.registeredObjects.filter { !$0.isFault && $0 is Self }.filter { predicate.evaluate(with: $0) }.compactMap { $0 as? Self }
-
+    
     return results
   }
 }
@@ -327,20 +327,20 @@ extension NSFetchRequestResult where Self: NSManagedObject {
   ///   - configuration: Configurable fetch request.
   /// - Returns: A cached object (if any).
   /// - Throws: It throws an error in cases of failure.
-  @available(iOS 10, tvOS 10, watchOS 3, macOS 10.12, *)
+  @available(iOS 10.0, tvOS 10.0, watchOS 3.0, macOS 10.12, *)
   public static func fetchCachedObject(in context: NSManagedObjectContext, forKey cacheKey: String, orCacheUsing configuration: @escaping (NSFetchRequest<Self>) -> Void) throws -> Self? {
     guard let cached = context.cachedManagedObject(forKey: cacheKey) as? Self else {
       let result = try fetchUniqueObject(in: context, with: configuration)
       context.setCachedManagedObject(result, forKey: cacheKey)
-
+      
       return result
     }
-
+    
     return cached
   }
 }
 
-// MARK: - Batch Delete
+// MARK: - Batch Operations
 
 extension NSFetchRequestResult where Self: NSManagedObject {
   /// **CoreDataPlus**
@@ -353,15 +353,18 @@ extension NSFetchRequestResult where Self: NSManagedObject {
   /// - Throws: It throws an error in cases of failure.
   /// - Note: A batch delete can **only** be done on a SQLite store.
   public static func batchUpdateObjects(with context: NSManagedObjectContext,
-                                        configuration: ((NSBatchUpdateRequest) -> Void)? = nil) throws -> NSBatchUpdateResult {
+                                        resultType: NSBatchUpdateRequestResultType = .statusOnlyResultType,
+                                        propertiesToUpdate: [AnyHashable : Any],
+                                        includesSubentities: Bool = true,
+                                        predicate: NSPredicate? = nil) throws -> NSBatchUpdateResult {
     guard context.persistentStoreCoordinator != nil else { throw NSError.persistentStoreCoordinatorNotFound(context: context) }
-
+    
     let batchRequest = NSBatchUpdateRequest(entityName: entityName)
-    configuration?(batchRequest)
-    /// propertiesToUpdate: Dictionary of NSPropertyDescription|property name string -> constantValue/NSExpression pairs describing the desired updates.
-    /// The expressions can be any NSExpression that evaluates to a scalar value.
-    /// Instead of nil use: NSExpression(forConstantValue: nil)
-
+    batchRequest.resultType = resultType
+    batchRequest.propertiesToUpdate = propertiesToUpdate
+    batchRequest.includesSubentities = includesSubentities
+    batchRequest.predicate = predicate
+    
     do {
       // swiftlint:disable:next force_cast
       return try context.execute(batchRequest) as! NSBatchUpdateResult
@@ -369,7 +372,7 @@ extension NSFetchRequestResult where Self: NSManagedObject {
       throw NSError.batchUpdateFailed(underlyingError: error)
     }
   }
-
+  
   /// **CoreDataPlus**
   ///
   /// Executes a batch delete on the context's persistent store coordinator.
@@ -385,19 +388,46 @@ extension NSFetchRequestResult where Self: NSManagedObject {
                                         resultType: NSBatchDeleteRequestResultType = .resultTypeStatusOnly,
                                         configuration: ((NSFetchRequest<Self>) -> Void)? = nil) throws -> NSBatchDeleteResult {
     guard context.persistentStoreCoordinator != nil else { throw NSError.persistentStoreCoordinatorNotFound(context: context) }
-
+    
     let request = NSFetchRequest<Self>(entityName: entityName)
     configuration?(request)
-
+    
     // swiftlint:disable:next force_cast
     let batchRequest = NSBatchDeleteRequest(fetchRequest: request as! NSFetchRequest<NSFetchRequestResult>)
     batchRequest.resultType = resultType
-
+    
     do {
       // swiftlint:disable:next force_cast
       return try context.execute(batchRequest) as! NSBatchDeleteResult
     } catch {
       throw NSError.batchDeleteFailed(underlyingError: error)
+    }
+  }
+  
+  /// **CoreDataPlus**
+  ///
+  /// Executes a batch insert on the context's persistent store coordinator.
+  /// - Parameters:
+  ///   - context: The context whose the persistent store coordinator will be used to execute the batch insert.
+  ///   - resultType: The type of the batch insert result (default: `NSBatchInsertRequestResultType.statusOnly`).
+  ///   - objects: A dictionary of objects to insert.
+  /// - Returns: a NSBatchInsertResult result.
+  /// - Throws: It throws an error in cases of failure.
+  /// - Note: A batch insert can **only** be done on a SQLite store.
+  @available(iOS 13.0, iOSApplicationExtension 13.0, tvOS 13.0, watchOS 6.0, macOS 10.15, *)
+  public static func batchInsertObjects(with context: NSManagedObjectContext,
+                                        resultType: NSBatchInsertRequestResultType = .statusOnly,
+                                        objects: [[String: Any]]) throws -> NSBatchInsertResult {
+    guard context.persistentStoreCoordinator != nil else { throw NSError.persistentStoreCoordinatorNotFound(context: context) }
+    
+    let batchRequest = NSBatchInsertRequest(entityName: entityName, objects: objects)
+    batchRequest.resultType = resultType
+    
+    do {
+      // swiftlint:disable:next force_cast
+      return try context.execute(batchRequest) as! NSBatchInsertResult
+    } catch {
+      throw NSError.batchInsertFailed(underlyingError: error)
     }
   }
 }